pyproject.toml
[tool.poetry]
name = "bml-ocr"
version = "0.1.0"
description = "A package to extract details from BML transfer receipts."
authors = ["Naffah Abdulla Rasheed <mail@naffah.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchaudio = "^2.3.0"
easyocr = "^1.7.1"
python-levenshtein = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
